Add JS createNewBoard() to POST a request to /add-board
The function creates a JSON dictionary transmitted over HTTP (hence the body as
a string and headers to specify the Content-Type) to /add-board.  Afterwards it
calls the callback function with the response from the server as argument.

This may seem redundant, but if server was not localhost and the client and
server were on different machines, this function takes a variable (boardTitle)
created client-side and tells the database (which is server-side) to update
with that information.