Update showBoards to use new containers; Add eventListener for button
The eventListener is added in the init() function (as it should be done
once) The function to be executed on the click is to call a prompt to
input the boardTitle that is used in dataHandler.createNewBoard(). That,
in turn, sends the POST to the server.  On a response from the server it
should update the DOM with the new board. This is currently done via a
page refresh ( location.reload(), which essentially gets all the boards
again from server and rebuilds the DOM again via showBoards) Something
like appending a child (without a page refresh and querying the server
DB) could work in the future.

The prompt may be changed to something like a modal (especially if using
Bootstrap) in the future.