Add board deletion (DOM only)
The board delete button (.remove-board) is now added dynamically (placed
inside .board-header div-container) Thus, .board-header is now a flex
row.  Inside showBoards, eventListeners are added for each such button,
and also boards (the .board div-container) are created with
data-id-of-board (since we need IDs for board deletion server-side).
These changes are also reflected when creating a newChildBoard
(data-id-... for .board, and eventListener for .remove-board).

Deletion is fairly simple in DOM, since children of the board
(boardNode) are also deleted when a DOM node (HTML element) is removed.

Also, a corresponding dataHandler.deleteBoard(boardID) which is supposed
to POST this delete request to server has been added for the sake of
completeness, but it does nothing currently (empty function).