Add connection*.py to handle SQL queries
Needs psycopg2 (reflected in updated 'requirements.txt').
Remember to export the environment variables. (see connection_wrapper)

The connection handler wrapper is in 'connection_wrapper.py' to
separate it from the actual SQL queries (in 'connection.py').
The queries are very general and perhaps obfuscated by the need to build
a list of SQL identifiers, literals and values (we use psycopg2.sql to
safely pass variables dynamically).  The actual functionality of each
query is visible in the multi-line strings (""" """). Specific
instances for these functions (such as 'get_boards()') will be made
available in data_handler.py in the next commit.

Ported these general functions from previous Ask-Mate project so some
previous influence remains (we use column_names_dict as a constant to
get all the table headers; we can update/write partial information to
tables by passing a dictionary and using intersection_fields to
determine which keys in the dictionary match which column names of the
table)

Kept 'persistence.py' for archival reasons, but it will no longer be
imported by data_handler.py