Update data_handler to support connection.py
Simplified some functions and moved the generic persistence. ones (such
as get_statuses to the end of this 'data_handler.py').
Also changed 'main.py' to account for the fact we no longer have a list,
but a dict with named keys.  Maybe this is why some of the code is still
buggy (cards not showing up on boards)

Left comments with TODOs for changes.  Not fully tested.